// src/types/event.type.ts
export enum Event {
  FAILURE = 'FAILURE',
  SUCCESS = 'SUCCESS',
  CONNECTION = 'CONNECTION',
  FRIEND_STATUS = 'FRIEND_STATUS',
  FRIEND_LIST = 'FRIEND_LIST',
  PRIVATE_MESSAGE = 'PRIVATE_MESSAGE',
  GROUP_MESSAGE = 'GROUP_MESSAGE',
  ROOM_CREATION = 'ROOM_CREATION',
  ROOM_JOINING = 'ROOM_JOINING',
  ROOM_LEAVING = 'ROOM_LEAVING',
  ROOM_JOINING_BROADCASTING = 'ROOM_JOINING_BROADCASTING',
  ROOM_MESSAGE = 'ROOM_MESSAGE',
}